testcase/le_tc/kernel: Move libc mqueue testcases to tc_mqueue file.
This patch shifts mq_setattr() and mq_getattr() testcases to
tc_mqueue.c and removes tc_libc_mqueue.c file.
